+++ packages/mobx-state-tree/__tests__/core/reference-onInvalidated.test.ts
@@ -392,23 +392,6 @@ describe("safeReferenceForCollection", () => {
 
     const SafeRef = types.safeReferenceForCollection(MyRefModel)
 
-    it("fails when a model property is invalidated", () => {
-        const Store = types.model({
-            todos: types.array(MyRefModel),
-            single: SafeRef
-        })
-
-        const store = Store.create({
-            todos: [{ id: "1" }, { id: "2" }],
-            single: "1"
-        })
-        unprotect(store)
-
-        expect(() => {
-            store.todos.splice(0, 1)
-        }).toThrow("value `undefined` is not assignable to type")
-    })
-
     it("removes invalidates items from map/array", () => {
         const Store = types.model({
             todos: types.array(MyRefModel),
@@ -431,34 +414,53 @@ describe("safeReferenceForCollection", () => {
         expect(store.map.size).toBe(1)
     })
 
-    it("does not accept undefined in the array", () => {
-        const Store = types.model({
-            todos: types.array(MyRefModel),
-            arr: types.array(SafeRef)
+    if (process.env.NODE_ENV !== "production") {
+        it("throws when a model property is invalidated", () => {
+            const Store = types.model({
+                todos: types.array(MyRefModel),
+                single: SafeRef
+            })
+
+            const store = Store.create({
+                todos: [{ id: "1" }, { id: "2" }],
+                single: "1"
+            })
+            unprotect(store)
+
+            expect(() => {
+                store.todos.splice(0, 1)
+            }).toThrow("value `undefined` is not assignable to type")
         })
 
-        expect(() =>
-            Store.create({
-                todos: [{ id: "1" }, { id: "2" }],
-                arr: ["1", undefined as any]
+        it("does not accept undefined in the array", () => {
+            const Store = types.model({
+                todos: types.array(MyRefModel),
+                arr: types.array(SafeRef)
             })
-        ).toThrow("value `undefined` is not assignable to type")
-    })
 
-    it("does not accept undefined in the map", () => {
-        const Store = types.model({
-            todos: types.array(MyRefModel),
-            map: types.map(SafeRef)
+            expect(() =>
+                Store.create({
+                    todos: [{ id: "1" }, { id: "2" }],
+                    arr: ["1", undefined as any]
+                })
+            ).toThrow("value `undefined` is not assignable to type")
         })
 
-        expect(() =>
-            Store.create({
-                todos: [{ id: "1" }, { id: "2" }],
-                map: {
-                    a1: "1",
-                    a2: undefined as any
-                }
+        it("does not accept undefined in the map", () => {
+            const Store = types.model({
+                todos: types.array(MyRefModel),
+                map: types.map(SafeRef)
             })
-        ).toThrow("value `undefined` is not assignable to type")
-    })
+
+            expect(() =>
+                Store.create({
+                    todos: [{ id: "1" }, { id: "2" }],
+                    map: {
+                        a1: "1",
+                        a2: undefined as any
+                    }
+                })
+            ).toThrow("value `undefined` is not assignable to type")
+        })
+    }
 })